Start skills page by adding tile styles and making them an option within the accordi on component.
Create tiles SCSS partial for handling the layout of the tiles on the skills page.

Create tile SCSS partial for individually styling a tile and its components.

Import both in the /components/_all.scss.

Add tileIcon prop to the Accordion component to determine if the accordion should output like a tile.

Update AccordionGroup component to pass the tileIcon prop when creating child Accordions.