Split Accordion component into two separate components: Disclosure and Accordions.
Setup new Disclosure and Accordions elements to automatically determine which children elements act as the button expanded/collapasing and which children are expanded/collapsed.

Create separate scroll utility used to scroll to an element on the page if it is not already visible. Implement scroll utility on Disclosure and Accordions element.

Update index and skills pages to use the new Disclosure and Accordions components.

Create new skills pattern for displaying relevant sub-skills of an overall skill.